+++ test/UnitTests/Caravel.Core.UnitTests/Tests/NodeAction/Atomic/AllowsToMergeLocalAndJourneyCancellationTokens.cs
@@ -1,6 +1,6 @@
 ﻿using AwesomeAssertions;
 
-namespace Caravel.Core.UnitTests.Tests.NodeAction.BetweenNavigations;
+namespace Caravel.Core.UnitTests.Tests.NodeAction.Atomic;
 
 [Trait(TestType, Unit)]
 [Trait(Feature, FeatureNodeAction)]
@@ -28,9 +28,6 @@ public sealed class AllowsToMergeLocalAndJourneyCancellationTokens : IDisposable
             .WithEdge<NodeSpy2>()
             .Done()
             .AddNode<NodeSpy2>()
-            .WithEdge<NodeSpy3>()
-            .Done()
-            .AddNode<NodeSpy3>()
             .Done();
 
         var journey = builder.Build();
@@ -38,8 +35,7 @@ public sealed class AllowsToMergeLocalAndJourneyCancellationTokens : IDisposable
         // Act
         var sut = async () =>
             await journey
-                .GotoAsync<NodeSpy2>()
-                .DoAsync<NodeSpy2>(
+                .DoAsync<NodeSpy1>(
                     async (node, ct) =>
                     {
                         // cancel only local token
@@ -48,7 +44,7 @@ public sealed class AllowsToMergeLocalAndJourneyCancellationTokens : IDisposable
                     },
                     _localTokenSource30mins.Token
                 )
-                .GotoAsync<NodeSpy3>();
+                .GotoAsync<NodeSpy2>();
 
         // Assert
         await sut.Should().ThrowExactlyAsync<OperationCanceledException>();
@@ -56,7 +52,7 @@ public sealed class AllowsToMergeLocalAndJourneyCancellationTokens : IDisposable
         _localTokenSource30mins.IsCancellationRequested.Should().BeTrue();
     }
 
-    [Fact(DisplayName = "When journey CancellationToken is set and local is not")]
+    [Fact(DisplayName = "When journey CancellationToken is default and local is set")]
     public async Task Test2()
     {
         // Arrange
@@ -65,9 +61,6 @@ public sealed class AllowsToMergeLocalAndJourneyCancellationTokens : IDisposable
             .WithEdge<NodeSpy2>()
             .Done()
             .AddNode<NodeSpy2>()
-            .WithEdge<NodeSpy3>()
-            .Done()
-            .AddNode<NodeSpy3>()
             .Done();
 
         using var journeyTokenSource = new CancellationTokenSource();
@@ -76,8 +69,7 @@ public sealed class AllowsToMergeLocalAndJourneyCancellationTokens : IDisposable
         // Act
         var sut = async () =>
             await journey
-                .GotoAsync<NodeSpy2>()
-                .DoAsync<NodeSpy2>(
+                .DoAsync<NodeSpy1>(
                     async (node, ct) =>
                     {
                         // cancel only journey token
@@ -85,7 +77,7 @@ public sealed class AllowsToMergeLocalAndJourneyCancellationTokens : IDisposable
                         return node;
                     }
                 )
-                .GotoAsync<NodeSpy3>();
+                .GotoAsync<NodeSpy2>();
 
         // Assert
         await sut.Should().ThrowExactlyAsync<OperationCanceledException>();
@@ -102,18 +94,14 @@ public sealed class AllowsToMergeLocalAndJourneyCancellationTokens : IDisposable
             .WithEdge<NodeSpy2>()
             .Done()
             .AddNode<NodeSpy2>()
-            .WithEdge<NodeSpy3>()
-            .Done()
-            .AddNode<NodeSpy3>()
             .Done();
 
         var journey = builder.Build(ct: CancellationToken.None);
 
         // Act
         var sut = await journey
-            .GotoAsync<NodeSpy2>()
-            .DoAsync<NodeSpy2>((node, ct) => Task.FromResult(node), CancellationToken.None)
-            .GotoAsync<NodeSpy3>();
+            .DoAsync<NodeSpy1>((node, ct) => Task.FromResult(node), CancellationToken.None)
+            .GotoAsync<NodeSpy2>();
 
         // Assert
         journey.JourneyCancellationToken.IsCancellationRequested.Should().BeFalse();
