+++ test/UnitTests/Caravel.Core.UnitTests/Tests/NodeAction/Atomic/ChangeTheCurrentNode_Snapshots/Test1.When_action_returns_a_different_node.verified.mmd
