+++ test/UnitTests/Caravel.Core.UnitTests/Tests/NodeAction/Atomic/ChangeTheCurrentNode_Snapshots/Test2.When_action_(with_journey)_returns_a_different_node.verified.mmd
