+++ test/UnitTests/Caravel.Core.UnitTests/Tests/NodeAction/Atomic/DoesNotChangeTheCurrentNode_Snapshots/Test2.When_action_(with_journey)_is_done_on_current_node.verified.mmd
