+++ test/UnitTests/Caravel.Core.UnitTests/Tests/NodeAction/Atomic/HasDefaultJourneyAndLocalCancellationTokensToNone_Snapshots/Test1.When_journey_CancellationToken_and_local_are_default.verified.mmd
