+++ test/UnitTests/Caravel.Core.UnitTests/Tests/NodeAction/Chained/AllowsToMergeLocalAndJourneyCancellationTokens_Snapshots/Test3.When_journey_CancellationToken_and_local_are_CancellationToken.None.verified.mmd
