+++ test/UnitTests/Caravel.Core.UnitTests/Tests/NodeAction/Chained/DoesNotChangeTheCurrentNode_Snapshots/Test1.When_action_is_done_on_current_node.verified.mmd
