+++ test/UnitTests/Caravel.Core.UnitTests/Tests/NodeAction/Chained/HasDefaultJourneyAndLocalCancellationTokensToNone.cs
@@ -1,6 +1,6 @@
 ﻿using AwesomeAssertions;
 
-namespace Caravel.Core.UnitTests.Tests.NodeAction.BeforeAnyNavigation;
+namespace Caravel.Core.UnitTests.Tests.NodeAction.Chained;
 
 [Trait(TestType, Unit)]
 [Trait(Feature, FeatureNodeAction)]
@@ -28,6 +28,9 @@ public sealed class HasDefaultJourneyAndLocalCancellationTokensToNone : IDisposa
             .WithEdge<NodeSpy2>()
             .Done()
             .AddNode<NodeSpy2>()
+            .WithEdge<NodeSpy3>()
+            .Done()
+            .AddNode<NodeSpy3>()
             .Done();
 
         // CancellationToken not set
@@ -35,8 +38,9 @@ public sealed class HasDefaultJourneyAndLocalCancellationTokensToNone : IDisposa
 
         // Act
         var sut = await journey
-            .DoAsync<NodeSpy1>((node, ct) => Task.FromResult(node))
-            .GotoAsync<NodeSpy2>(); // CancellationToken not set
+            .GotoAsync<NodeSpy2>()
+            .DoAsync<NodeSpy2>((node, ct) => Task.FromResult(node)) // CancellationToken not set
+            .GotoAsync<NodeSpy3>();
 
         // Assert
         journey.JourneyCancellationToken.IsCancellationRequested.Should().BeFalse();
